品类管理模快

+++ src/components/header/header.jsx
@@ -1,5 +1,6 @@
 import React , {Component} from 'react'
 import {withRouter} from 'react-router-dom'
+import { Modal} from 'antd'
 import {formatTime} from '../../utils/utils'
 import './header.less'
 import menuList from '../../config/menuConfig'
@@ -9,12 +10,23 @@ class Header extends Component {
    }
    //退出
    loginOut = () => {
-       localStorage.clear()
-       this.props.history.push('/login')
+    Modal.confirm({
+        title: '提示',
+        content: '确定退出吗',
+        okText: '确定',
+        cancelText: '取消',
+        onOk: () => {
+          localStorage.clear()
+          this.props.history.push('/login')
+        },
+        onCancel() {
+          console.log('Cancel');
+        },
+      });
+      
    }
    //每次切换路由的的时候切换标题
    getTitle = (menuList) => {
-       console.log('getTitle')
        //获取当前路由的路径 注意 header组件并不是路由组件 是通过withRoute包装后才能获取路由属性
        const path = this.props.location.pathname;
        let title;
@@ -50,7 +62,6 @@ class Header extends Component {
         const {currentDate} = this.state
         //每次render的时候切换标题
         const title = this.getTitle(menuList)
-        console.log('getTitle',title);
         return (
             <div className='headerContainer'>
                 <div className="headerContainerTop">
